pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "fuzzywuzzy>=0.18.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "pytesseract>=0.3.13",
    "python-levenshtein>=0.27.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.3",
]
